LayoutTests/storage/indexeddb/versionchangerequest-activedomobject.html

<!DOCTYPE html>
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Ensure that IDBVersionChangeRequest objects are not GC'd if they have pending events");

function test() {
    evalAndLog("indexedDB = window.indexedDB || window.webkitIndexedDB || window.mozIndexedDB;");
    shouldBeFalse("indexedDB == null");

    debug("");
    evalAndLog("window.dbname = 'versionchangerequest-activedomobject'");

    testDeleteDatabase();
}

function testDeleteDatabase()
{
    debug("");
    debug("testDeleteDatabase():");
    // NOTE: deleteRequest is local variable so it is not captured in global scope.
    var deleteRequest = evalAndLog("indexedDB.deleteDatabase(window.dbname)");
    deleteRequest.onerror = unexpectedErrorCallback;
    deleteRequest.onblocked = unexpectedBlockedCallback;
    deleteRequest.onsuccess = function() { deleteSuccess(); };
    deleteRequest = null;
    if (window.gc) {
        evalAndLog("window.gc()");
    }
}

function deleteSuccess()
{
    testPassed("deleteDatabase's IDBVersionChangeRequest.onsuccess event fired");
    testSetVersion();
}

function testSetVersion()
{
    debug("");
    debug("testSetVersion():");

    // NOTE: This is just an IDBRequest, it can be global.
    evalAndLog("openRequest = indexedDB.open(window.dbname)");
    openRequest.onerror = unexpectedErrorCallback;
    openRequest.onsuccess = function () {
        evalAndLog("db = openRequest.result");

        // NOTE: versionRequest is local variable so it is not captured in global scope.
        var versionRequest = evalAndLog("db.setVersion('1')");
        versionRequest.onerror = unexpectedErrorCallback;
        versionRequest.onblocked = unexpectedBlockedCallback;
        versionRequest.onsuccess = function() { versionSuccess(); };
        versionRequest = null;
        if (window.gc) {
            evalAndLog("window.gc()");
        }
    };
}

function versionSuccess()
{
    testPassed("setVersion's IDBVersionChangeRequest.onsuccess event fired");
    finishJSTest();
}

test();

</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
